requirements.txt
aioboto3==12.1.0
aiogoogle==5.1.0
aiohttp-retry==2.8.3
aiomysql==0.2.0
antlr4-python3-runtime==4.8
asyncio==3.4.3
asynctest==0.13.0
attrs==23.1.0
azure-identity==1.16.1
azure-monitor-query==1.0.2
colorlog==6.8.2
datadog_api_client[async]==2.12.0
flask==3.0.0
flatten_json==0.1.14
json-fix==1.0.0
jsonmerge==1.9.2
jsonref==1.1.0
mysql-connector-python==8.0.25
numpy==1.24.4
onelogin==2.0.4
pandas==1.5.2
pyOpenSSL==24.1.0
pyparsing==3.0.9
python-dateutil==2.8.2
regex==2023.12.25
stix2-matcher==3.0.0
stix2-patterns==1.3.2
sumologic-sdk==0.1.13
urllib3==1.26.19
xmltodict==0.13.0
